refactor(components.notification): move to dedicatedUniverseComponent
