+++ client/app/dedicatedUniverseComponents/index.js
@@ -1,9 +1,12 @@
 import angular from 'angular';
 
+import ducNotification from './notification';
+
 const moduleName = 'dedicatedUniverseComponents';
 
 angular
   .module(moduleName, [
+    ducNotification,
   ]);
 
 export default moduleName;
